Move edge resolvers to their own file.
